# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ria Noel Shop Invoice API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [passcode]
              properties:
                passcode: { type: string }
      responses:
        '200': { description: Authenticated }
        '401': { description: Invalid credentials }
...